+++ CHANGELOG.md
@@ -6,11 +6,21 @@ All notable changes to this project are documented in this file.
 
 ### Added
 
-- No improvements released yet
+- Nothing
 
 ### Fixed
 
-- No fixes released yet.
+- Nothing
+
+## [1.1.0] - 2026-09-19
+
+### Changes
+
+- Add Hugging Face integration
+- Experiment accessibility
+- Automatic experiment detection in visualizer
+- Optimizer LR tracking during training
+- Add new callback: ReduceLROnPlateauCallback
 
 ## [1.0.0] - 2026-09-17